refactor : view를 이벤트 단위별로 재구성
