--- src/main/java/racingcar/view/OutputView.java
@@ -1,7 +1,0 @@
-package racingcar.view;
-
-public class OutputView {
-    public static void printMessage(String message) {
-        System.out.println(message);
-    }
-}
